swipe to delete, dialog to ask if item should be deleted

+++ app/src/main/java/com/example/myshop/ui/adapters/AddressAdapter.kt
@@ -1,6 +1,7 @@
 package com.example.myshop.ui.adapters
 
 import android.app.Activity
+import android.app.AlertDialog
 import android.content.Context
 import android.content.Intent
 import android.view.LayoutInflater
@@ -41,10 +42,26 @@ class AddressAdapter( private val context: Context, private val addressList: Arr
         notifyItemChanged(position)
     }
     fun notifyDeleteItem(activity: AddressListActivity, position: Int){
-        activity.showDialogProgress(context.resources.getString(R.string.please_wait))
-        FirestoreClass().deleteAddress(activity, addressList[position].id)
-        notifyItemChanged(position)
-        FirestoreClass().getAddressList(activity)
+
+            val builder= AlertDialog.Builder(context)
+            builder.setTitle("Delete")
+            builder.setMessage("Are you sure you want to delete this Address?")
+            builder.setIcon(android.R.drawable.ic_dialog_alert)
+            builder.setPositiveButton("Yes"){ dialogInterface, _ ->
+                //delete the address
+                activity.showDialogProgress(context.resources.getString(R.string.please_wait))
+                FirestoreClass().deleteAddress(activity, addressList[position].id)
+                notifyItemChanged(position)
+                FirestoreClass().getAddressList(activity)
+                dialogInterface.dismiss()
+            }
+            builder.setNegativeButton("No"){ dialogInterface, _ ->
+                notifyItemChanged(position)
+                dialogInterface.dismiss()
+            }
+            val alertDialog: AlertDialog = builder.create()
+            alertDialog.setCancelable(false)
+            alertDialog.show()
     }
 
     override fun getItemCount(): Int {
